use separate datasource for integration tests
