+++ src/types/datasource.d.ts
@@ -1,0 +1,3 @@
+import { DataSource } from 'typeorm';
+
+export type IDataSource = DataSource | 'Other'; //Add any other data source type using | as needed
